requirements.txt
# This project requires Python 3.13
Appium-Python-Client>=5.1.2
selenium>=4.34.2
playwright>=1.54.0
pytest>=8.4.1
pytest-html>=4.1.1
psutil>=7.0.0